requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg[binary]==3.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
requests==2.31.0

